[Silver I] Title: 오르막 수, Time: 80 ms, Memory: 11848 KB -BaekjoonHub

+++ 백준/Silver/11057. 오르막 수/오르막 수.java
@@ -2,8 +2,8 @@ import java.io.*;
 public class Main {
     static BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
     static BufferedWriter bw = new BufferedWriter(new OutputStreamWriter(System.out));
-    static final long MOD = 10007;
-    static long[][] mat = new long[10][10];
+    static final int MOD = 10007;
+    static int[][] mat = new int[10][10];
     public static void main(String[] args) throws Exception {
         for(int r=0;r<10;r++){
             for(int c=0;c<10;c++){
@@ -15,8 +15,8 @@ public class Main {
         bw.write(String.format("%d",matPow(mat,N+1)[0][9]));
         bw.flush();
     }
-    static long[][] matMul(long[][] a, long[][] b){
-        long[][] c = new long[b[0].length][a.length];
+    static int[][] matMul(int[][] a, int[][] b){
+        int[][] c = new int[b[0].length][a.length];
         for(int i=0;i<a.length;i++){
             for(int j=0;j<a[0].length;j++){
                 for(int k=0;k<b[0].length;k++){
@@ -26,7 +26,7 @@ public class Main {
         }
         return c;
     }
-    static long[][] matPow(long[][] a, long num){
+    static int[][] matPow(int[][] a, int num){
         if(num==1) return a;
         else if(num%2==1) return matMul(matPow(a,num-1),a);
         else return matPow(matMul(a,a),num/2);
